fallback-fastapi/requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
alembic==1.8.0
anyio==3.6.1
APScheduler==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
azure-ai-formrecognizer==3.2.0b2
azure-common==1.1.28
azure-core==1.22.1
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.11.1
bidict==0.22.0
bleach==5.0.1
boto3==1.24.20
botocore==1.27.20
CacheControl==0.12.11
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==1.6.0
colorama==0.4.5
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==37.0.2
cycler==0.11.0
dask==2022.2.0
databricks-cli==0.17.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
dnspython==1.16.0
docker==5.0.3
docopt==0.6.2
elastic-apm==6.10.1
elasticsearch==7.10.1
entrypoints==0.4
executing==0.8.3
farm-haystack @ git+https://github.com/deepset-ai/haystack.git@ba08fc86f5871555537cc3766889e9cae8c5ad03
fastapi==0.78.0
fastjsonschema==2.15.3
fbmessenger==6.0.0
filelock==3.7.1
fire==0.4.0
Flask==2.1.2
flatbuffers==2.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h11==0.13.0
h5py==3.7.0
httptools==0.4.0
huggingface-hub==0.7.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
inflect==5.6.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.1.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
langdetect==1.0.9
libclang==14.0.1
locket==1.0.0
lxml==4.9.1
Mako==1.2.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mattermostwrapper==2.2
mistune==0.8.4
mlflow==1.27.0
mmh3==3.0.0
monotonic==1.6
more-itertools==8.13.0
msgpack==1.0.4
msrest==0.6.21
multidict==5.2.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
nltk==3.7
notebook==6.4.12
num2words==0.5.10
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pdf2image==1.14.0
pickleshare==0.7.5
Pillow==9.1.1
posthog==1.4.9
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
prompt-toolkit==2.0.10
protobuf==3.19.4
psutil==5.9.1
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pydot==1.4.2
Pygments==2.12.0
PyJWT==2.4.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.18.1
pyspellchecker==0.6.3
pyTelegramBotAPI==4.6.0
pytesseract==0.3.7
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2021.3
pywin32==227
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
quantulum3==0.7.10
querystring-parser==1.2.4
questionary==1.10.0
randomname==0.1.5
rapidfuzz==2.1.3
rasa==3.2.1
rasa-sdk==3.2.0
redis==3.5.3
regex==2022.4.24
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rocketchat-API==1.25.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sacremoses==0.0.53
sanic==21.12.1
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.1
scipy==1.7.3
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
sentry-sdk==1.3.1
seqeval==1.2.2
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.19.1
tabulate==0.8.10
tarsafe==0.0.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.3
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tika==1.24
tinycss2==1.1.1
tokenizers==0.12.1
toolz==0.11.2
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
twilio==6.50.1
typeguard==2.13.3
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzlocal==2.1
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.18.2
waitress==2.1.2
watchfiles==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
